+++ X8_multirotor_master/X8_multirotor/ip/Flight_Main/solution1/.autopilot/db/flightMain.pp.0.cpp
@@ -24667,7 +24667,7 @@ void flightmain (F16_t rcCmdIn[6], F16_t obj_avd_cmd[5], F16_t cmdOut[4096])
 
 
 
- isArmed = rcCmdIn[4];
+ isArmed = uint8_t(rcCmdIn[4]);
 
     flightModeFlag = uint8_t(rcCmdIn[5]);
